// src/app/profile/interfaces/profile.interface.ts
export interface Profile {
  name: string;
  title: string;
  dateBirth: string;
  ci: string;
  phone: string;
  email: string;
  direction: string;
  img: string;
  nationality: string;
}

export interface Education {
  name: string;
  date: string;
  location: string;
  title: string;
}

export interface Skills {
  name: string;
}